==> 1. Working with raw data.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_excel('/Users/elena/Documents/Диплом/Data/Maximum_Data_extract_20200229.xlsx')

df = df.replace('..', np.nan)

#List of variables for the research
series = ['Agricultural methane emissions (% of total)',
'Agricultural nitrous oxide emissions (% of total)',
'Air transport, freight (million ton-km)',
'Air transport, passengers carried',
'Air transport, registered carrier departures worldwide',
'Alternative and nuclear energy (% of total energy use)',
'Cereal yield (kg per hectare)',
'CO2 emissions (metric tons per capita)',
'CO2 emissions from electricity and heat production, total (% of total fuel combustion)',
'CO2 emissions from gaseous fuel consumption (% of total)',
'CO2 emissions from liquid fuel consumption (% of total)',
'CO2 emissions from manufacturing industries and construction (% of total fuel combustion)',
'CO2 emissions from other sectors, excluding residential buildings and commercial and public services (% of total fuel combustion)',
'CO2 emissions from residential buildings and commercial and public services (% of total fuel combustion)',
'CO2 emissions from solid fuel consumption (% of total)',
'CO2 emissions from transport (% of total fuel combustion)',
'CO2 intensity (kg per kg of oil equivalent energy use)',
'Electric power consumption (kWh per capita)',
'Electricity production from hydroelectric sources (% of total)',
'Electricity production from nuclear sources (% of total)',
'Electricity production from oil, gas and coal sources (% of total)',
'Electricity production from renewable sources, excluding hydroelectric (% of total)',
'Energy related methane emissions (% of total)',
'Energy use (kg of oil equivalent per capita)',
'Forest area (% of land area)',
'GDP per capita growth (annual %)',
'Land under cereal production (hectares)',
'Population growth (annual %)',
'Trade (% of GDP)']

df_v = df[df['Series Name'].isin(series)]

#List of countries for the reserch
countries = ['Australia',
'Austria',
'Belgium',
'Bulgaria',
'Canada',
'Croatia',
'Czech Republic',
'Denmark',
'Estonia',
'Euro area',
'European Union',
'Finland',
'France',
'Germany',
'Greece',
'Hungary',
'Iceland',
'Ireland',
'Italy',
'Japan',
'Latvia',
'Liechtenstein',
'Lithuania',
'Luxembourg',
'Monaco',
'Netherlands',
'New Zealand',
'Norway',
'Poland',
'Portugal',
'Romania',
'Russian Federation',
'Slovak Republic',
'Slovenia',
'Spain',
'Sweden',
'Switzerland',
'Ukraine',
'United Kingdom',
'United States']

df_vc = df_v[df['Country Name'].isin(countries)]

#Export the dataframe of variables for the chosen countries
df_vc.to_excel('Data_total.xlsx')

#Create a file for each variable
for item in series:
    df_vc[df['Series Name'] == item].to_excel('{}.xlsx'.format(item))
