==> 2. Collecting data by country.ipynb <==
import pandas as pd

df = pd.read_excel('Data_total.xlsx')

series = sorted(list(set(df['Series Name']))) #Collect names of varibles in the file

countries = set(df['Country Name']) 

#Create a file with one sheet for each country
with pd.ExcelWriter('Data_by_country.xlsx') as writer:  
    for item in countries:
        df[df['Country Name'] == item].to_excel(writer, sheet_name='{}'.format(item))
