==> 3. Collecting all variables in one file.ipynb <==
import pandas as pd
import numpy as np

import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

df = pd.read_excel('/Users/elena/Documents/Диплом/Data/Data_total.xlsx')

df_t = pd.read_excel('/Users/elena/Documents/Диплом/Data/av_temperature.xlsx')
df_t.index = list(range(1960, 2017))#year range

df_t = df_t.T

df_t['Series Name'] = 'Average annual temperature, C'
df_t['Series Code'] = 'av.an.temp'
#Now we have a dataframe of av.temp. for each country (row) by year(column)

country_id = df[['Country Name', 'Country Code']]
country_dict = country_id.set_index('Country Code').T.to_dict('list')

country_dict
