==> 4. Collecting global land temperature average by country.ipynb <==
import pandas as pd
import glob

for path in glob.glob('/Users/elena/Documents/Диплом/Data/температура по странам/raw data/*.csv'):
    df = pd.read_csv(path)
    df.to_csv('{}.csv'.format(path[-7:-4]), index=False)
#     print(df.columns)
#     df.index = range(0, 312)
#     print(path[-7:-4])
#     print(df.columns[1])
#     print(df.columns[0])
#     years = sorted(list(set(df[df.columns[1]])))
#     print(years)
#     c = []
#     for year in years:
#         c.append(df[df[df.columns[1]]==int(year)][df.columns[0]].mean())
#     print(path[-7:-4])
#     df_new = pd.DataFrame([years, c]).T
#     df_new.columns = ['t', 'AvT']
#     df_new.to_excel('{}.xlsx'.format(path[-7:-4]))

i=1
for path in glob.glob('/Users/elena/Documents/Диплом/Data/температура по странам/*.csv'):
    df = pd.read_csv(path)
    print(i, path[-7:-4])
    i += 1
    years = sorted(list(set(df[' Year'])))
    c = []
    for year in years:
        c.append(df[df[df.columns[1]]==int(year)][df.columns[0]].mean())
    df_new = pd.DataFrame([years, c]).T
    df_new.columns = ['t', 'AvT']
    df_new.to_excel('{}.xlsx'.format(path[-7:-4]))

years = list(range(1992,2015))
df_all = pd.DataFrame(years, columns=['t'])
c = []
names = []
for path in glob.glob('/Users/elena/Documents/Диплом/Data/температура по странам/*.xlsx'):
    #print(path[-8:-5])
    df_country = pd.read_excel(path)
    names.append(path[-8:-5])
    c.append(df_country[df_country['t'].isin(years)]['AvT'])

final_df = pd.DataFrame(c).T
final_df.columns = names
final_df.index=years
# final_df['t']=years

final_df.to_excel('/Users/elena/Documents/Диплом/Data/av_temperature.xlsx')

inputPath = '/Users/elena/Documents/Диплом/Data/температура по странам/usa and china/*.csv'

for path in glob.glob(inputPath):
    print(path)
    df = pd.read_csv(path)
    print(path[-7:-4])
    years = sorted(list(set(df[' Year'])))
    years = years[73:-2]
    c = []
    for year in years:
        c.append(df[df[df.columns[1]]==int(year)][df.columns[0]].mean())
    df_new = pd.DataFrame([years, c]).T
    df_new.columns = ['t', 'AvT']
    df_new.to_excel('{}_long.xlsx'.format(path[-7:-4]))

years[73:-2]
