==> 6. Descriptive statistics.ipynb <==
import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_excel('/Users/elena/Documents/Диплом/Data/Глава 1/описательные статистики.xlsx', sheet_name='Data mod')

df = df.drop(['Series Code', 'Среднее за период', 'Мин за период', 'Максимум за период'], axis=1)#remove redundant columns

#create a dataframe for each cluster
df_0 = df[df['Cluster'] == 1].drop('Cluster', axis=1)
df_1 = df[df['Cluster'] == 2].drop('Cluster', axis=1)
df_2 = df[df['Cluster'] == 3].drop('Cluster', axis=1)

#group country names by cluster
countries_0 = list(sorted(set(df_0['Country Code'])))
countries_1 = list(sorted(set(df_1['Country Code'])))
countries_2 = list(sorted(set(df_2['Country Code'])))

df_0[df_0[1992].isna() == True].head()

#Read temperature data from the file
temp = pd.read_excel('/Users/elena/Documents/Диплом/Data/av_temperature.xlsx')

temp = temp.rename(columns={'Unnamed: 0': 't'})

temp.index = temp.t
temp = temp.drop('t', axis=1)

temp = temp.T
temp['Series Name'] = 'Average Air temperature, Celsius'

temp['Country Code'] = temp.index

df = df.append(temp, sort=True)

series = list(set(df['Series Name']))

df[df['Country Code'] == 'Китай'].head()#Data for China

# Cluster of countries with 1-7% share of world CO2 emissions

c1_stats = pd.DataFrame([])
for country in countries_1:
    print(country)#to check the countries in this cluster
    df_country = df[df['Country Code'] == country]#dataframe for one country in the cluster
    df_country = df_country.drop(['Country Name', 'Country Code', 'Cluster'], axis=1)#remove redundant columns
    l = len(df_country)#number of variables for this country
    variable = []
    means = []
    variances = []
    mins = []
    maxs = []
    chng = []
    av_chng = []
    pct_chng = []
    for i in range(0, l):#for each variable
        variable.append(df_country.iloc[i,0])
        row = df_country.iloc[i, 1:].astype(float).dropna()
        if len(row) < 1:#take np.nan as descriptive statistics if there is no observations of the variable
            means.append(np.nan)
            variances.append(np.nan)
            mins.append(np.nan)
            maxs.append(np.nan)
            chng.append(np.nan)
            av_chng.append(np.nan)
            pct_chng.append(np.nan)
        else:
            means.append(row.mean())
            variances.append(row.std())
            mins.append(row.min())
            maxs.append(row.max())
            if row.iloc[0] == 0:
                chng.append(0)
            else:
                chng.append((row.iloc[-1]-row.iloc[0])/row.iloc[0]*100)
            av_chng.append(row.diff().mean())
            pct_chng.append(row.pct_change().mean())
    df_ = pd.DataFrame([variable, means, variances, mins, maxs, chng, av_chng, pct_chng]).T
    df_.columns = ['Переменная', 'Среднее значение', 'СКО', 'Минимальное значение', 'Максимальное значение', 'Изменение за период, %', 'Среднее изменение за период','Средний темп прироста, %']
    df_.to_excel('{}_DS.xlsx'.format(country))#save descriptive statistics for each country
    df_['Country Code'] = country
    c1_stats = pd.concat([c1_stats, df_])
c1_stats.to_excel('Cl_1_7_DS.xlsx')#save cluster stats

c1_stats.head()

agr_stats_1 = ['Среднее значение', 'СКО', 'Минимальное значение', 'Максимальное значение', 'Изменение за период, %', 'Среднее изменение за период','Средний темп прироста, %']
agr_stats_1_values = []

factors = ['Объем грузовых авиаперевозок, млн тонн-км', 'Объем пассажирских авиаперевозок, млн чел.', 'Выбросы углекислого газа от транспорта, доля от общего объема при сжигании топлива', 'Выбросы углекислого газа, кт на душу населения', 'Выбросы углекислого газа от производства и строительства, доля от общего объема при сжигании топлива', 'Урожайность круп, кг на гектар', 'Площадь лесных массивов, доля суши', 'Потребление электроэнергии (kWh per capita)', 'Реальный ВВП на душу населения, долл. США', 'Темп роста реального ВВП на душу населения в долл. США', 'Среднегодовая температура воздуха']

stat_1 = pd.DataFrame([])
for stat in agr_stats_1:
    stat_1_val = []
    if stat == 'Минимальное значение':
        for item in series:
            stat_1_val.append(c1_stats[c1_stats['Переменная']==item][stat].min())
            #print(min(c0_stats[c0_stats['Переменная']==item][stat]))
    elif stat == 'Максимальное значение':
        for item in series:
            stat_1_val.append(c1_stats[c1_stats['Переменная']==item][stat].max())
    else:
        for item in series:
            stat_1_val.append(c1_stats[c1_stats['Переменная']==item][stat].mean())
    stat_1[stat] = stat_1_val
stat_1.index = series
stat_1[stat_1.index.isin(factors)].to_excel('AG_cl_1_7.xlsx')

# Cluster of leaders on CO2 emissions

c0_stats = pd.DataFrame([])
for country in countries_0:
    print(country)
    df_country = df[df['Country Code'] == country]
    df_country = df_country.drop(['Country Name', 'Country Code', 'Cluster'], axis=1)
    #print(df_country.columns)
    l = len(df_country)
    variable = []
    means = []
    variances = []
    mins = []
    maxs = []
    chng = []
    av_chng = []
    pct_chng = []
    for i in range(0, l):
        variable.append(df_country.iloc[i,0])
        row = df_country.iloc[i, 1:].astype(float).dropna()
        if len(row) < 1:
            means.append(np.nan)
            variances.append(np.nan)
            mins.append(np.nan)
            maxs.append(np.nan)
            chng.append(np.nan)
            av_chng.append(np.nan)
            pct_chng.append(np.nan)
        else:
            means.append(row.mean())
            variances.append(row.std())
            mins.append(row.min())
            maxs.append(row.max())
            if row.iloc[0] == 0:
                chng.append(0)
            else:
                chng.append((row.iloc[-1]-row.iloc[0])/row.iloc[0]*100)
            av_chng.append(row.diff().mean())
            pct_chng.append(row.pct_change().mean())
    df_ = pd.DataFrame([variable, means, variances, mins, maxs, chng, av_chng, pct_chng]).T
    df_.columns = ['Переменная', 'Среднее значение', 'СКО', 'Минимальное значение', 'Максимальное значение', 'Изменение за период, %', 'Среднее изменение за период','Средний темп прироста, %']
    df_.to_excel('{}_DS.xlsx'.format(country))
    df_['Country Code'] = country
    c0_stats = pd.concat([c0_stats, df_])
c0_stats.to_excel('Losers_DS.xlsx')

agr_stats_0 = ['Среднее значение', 'СКО', 'Минимальное значение', 'Максимальное значение', 'Изменение за период, %', 'Среднее изменение за период','Средний темп прироста, %']
agr_stats_0_values = []
stat_0 = pd.DataFrame([])
for stat in agr_stats_0:
    stat_0_val = []
    if stat == 'Минимальное значение':
        for item in series:
            stat_0_val.append(c0_stats[c0_stats['Переменная']==item][stat].min())
            #print(min(c0_stats[c0_stats['Переменная']==item][stat]))
    elif stat == 'Максимальное значение':
        for item in series:
            stat_0_val.append(c0_stats[c0_stats['Переменная']==item][stat].max())
    else:
        for item in series:
            stat_0_val.append(c0_stats[c0_stats['Переменная']==item][stat].mean())
    stat_0[stat] = stat_0_val
stat_0.index = series
stat_0[stat_0.index.isin(factors)].to_excel('AG_Leaders.xlsx')

# Cluster of countries with <1% share of world CO2 emissions

c2_stats = pd.DataFrame([])
for country in countries_2:
    print(country)
    df_country = df[df['Country Code'] == country]
    df_country = df_country.drop(['Country Name', 'Country Code', 'Cluster'], axis=1)
    l = len(df_country)
    variable = []
    means = []
    variances = []
    mins = []
    maxs = []
    chng = []
    av_chng = []
    pct_chng = []
    for i in range(0, l):
        variable.append(df_country.iloc[i,0])
        row = df_country.iloc[i, 1:].astype(float).dropna()
        if len(row) < 1:
            means.append(np.nan)
            variances.append(np.nan)
            mins.append(np.nan)
            maxs.append(np.nan)
            chng.append(np.nan)
            av_chng.append(np.nan)
            pct_chng.append(np.nan)
        else:
            means.append(row.mean())
            variances.append(row.std())
            mins.append(row.min())
            maxs.append(row.max())
            if row.iloc[0] == 0:
                chng.append(0)
            else:
                chng.append((row.iloc[-1]-row.iloc[0])/row.iloc[0]*100)
            av_chng.append(row.diff().mean())
            pct_chng.append(row.pct_change().mean())
    df_ = pd.DataFrame([variable, means, variances, mins, maxs, chng, av_chng, pct_chng]).T
    df_.columns = ['Переменная', 'Среднее значение', 'СКО', 'Минимальное значение', 'Максимальное значение', 'Изменение за период, %', 'Среднее изменение за период','Средний темп прироста, %']
    df_.to_excel('{}_DS.xlsx'.format(country))
    df_['Country Code'] = country
    c2_stats = pd.concat([c2_stats, df_])
c2_stats.to_excel('Middlers_DS.xlsx')

agr_stats_2 = ['Среднее значение', 'СКО', 'Минимальное значение', 'Максимальное значение', 'Изменение за период, %', 'Среднее изменение за период','Средний темп прироста, %']
agr_stats_2_values = []
stat_2 = pd.DataFrame([])
for stat in agr_stats_2:
    stat_2_val = []
    if stat == 'Минимальное значение':
        for item in series:
            stat_2_val.append(c2_stats[c2_stats['Переменная']==item][stat].min())
            #print(min(c0_stats[c0_stats['Переменная']==item][stat]))
    elif stat == 'Максимальное значение':
        for item in series:
            stat_2_val.append(c2_stats[c2_stats['Переменная']==item][stat].max())
    else:
        for item in series:
            stat_2_val.append(c2_stats[c2_stats['Переменная']==item][stat].mean())
    stat_2[stat] = stat_2_val
stat_2.index = series
stat_2[stat_2.index.isin(factors)].to_excel('AG_Middlers.xlsx')

temp.to_excel('temp_for_DS.xlsx')
